- Started 2D sprite rendering support - Added  caching of width and height of texture in texture class - Added texture manager to hold multiple textures - Added assert to simple draw in static initialize to check nullptr - Added Sprite types header to keep track of sprite types - Added Sprite renderer class to render sprites
git-tfs-id: [https://dev.azure.com/timjustin/]$/Angazi;C245

+++ Framework/Graphics/Inc/SimpleDraw.h
@@ -22,7 +22,7 @@ namespace Angazi::Graphics::SimpleDraw {
 	void AddTransform(const Math::Matrix4& transform);
 	void AddBone(const Math::Matrix4& transform);
 
-	void AddGroundPlane(float size, bool fill = true ,const Color& color = Colors::DarkGray);
+	void AddGroundPlane(int size, bool fill = true ,const Color& color = Colors::DarkGray);
 
 	//2D Simple Draw
 	void AddScreenLine(const Math::Vector2& v0, const Math::Vector2& v1, const Color& color);
